Muda nome projeto gateway para server-eureka, recipe para api-recipe e adiciona projeto gateway; Configuração do eureka e gateway
