--- server/api-eureka/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'api-eureka'
