--- server/api-eureka/src/main/java/com/gateway/apieureka/ApiEurekaApplication.java
@@ -1,15 +1,0 @@
-package com.gateway.apieureka;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
-
-@SpringBootApplication
-@EnableEurekaServer
-public class ApiEurekaApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ApiEurekaApplication.class, args);
-	}
-
-}
