--- server/api-eureka/src/main/resources/application.yml
@@ -1,11 +1,0 @@
-spring:
-  application:
-    name: api-gateway
-
-server:
-  port: 8761
-eureka:
-  client:
-    registerWithEureka: false
-    fetchRegistry: false
-
