--- server/api-eureka/src/test/java/com/gateway/apieureka/ApiEurekaApplicationTests.java
@@ -1,13 +1,0 @@
-package com.gateway.apieureka;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApiEurekaApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
