+++ server/api-gateway/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = 'api-gateway'
