+++ server/api-gateway/src/main/java/com/gateway/apigateway/ApiGatewayApplication.java
@@ -1,0 +1,13 @@
+package com.gateway.apigateway;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ApiGatewayApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ApiGatewayApplication.class, args);
+	}
+
+}
