+++ server/api-gateway/src/test/java/com/gateway/apigateway/ApiGatewayApplicationTests.java
@@ -1,0 +1,13 @@
+package com.gateway.apigateway;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApiGatewayApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
