+++ server/api-recipe/src/main/java/com/recipe/apirecipe/ApiRecipeApplication.java
@@ -1,0 +1,16 @@
+package com.recipe.apirecipe;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
+
+@SpringBootApplication
+@EnableDiscoveryClient
+public class ApiRecipeApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ApiRecipeApplication.class, args);
+	}
+
+}
