+++ server/api-recipe/src/main/resources/application.yml
@@ -1,0 +1,11 @@
+spring:
+  application:
+    name: api-recipe
+  liquibase:
+    contexts: dev
+    changeLog: classpath:db/changelog/master.xml
+
+eureka:
+  client:
+    serviceUrl:
+      defaultZone: ${eureka-host:http://localhost:8761/eureka/}
