--- server/recipe/src/main/java/com/recipe/app/AppApplication.java
@@ -1,16 +1,0 @@
-package com.recipe.app;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-
-@SpringBootApplication
-@EnableDiscoveryClient
-public class AppApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AppApplication.class, args);
-	}
-
-}
