+++ server/server-eureka/src/main/java/com/eureka/servereureka/ServerEurekaApplication.java
@@ -1,0 +1,15 @@
+package com.eureka.servereureka;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
+
+@SpringBootApplication
+@EnableEurekaServer
+public class ServerEurekaApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(com.eureka.servereureka.ServerEurekaApplication.class, args);
+	}
+
+}
