+++ server/server-eureka/src/main/resources/application.yml
@@ -1,0 +1,16 @@
+server:
+  port: 8761
+
+spring:
+  application:
+    name: server-eureka
+
+eureka:
+  instance:
+    hostname: localhost
+    preferIpAddress: false
+  client:
+    registerWithEureka: false
+    fetchRegistry: false
+    serviceUrl:
+      defaultZone: http://localhost:8761/eureka/
