+++ server/server-eureka/src/test/java/com/eureka/servereureka/ServerEurekaApplicationTests.java
@@ -1,0 +1,13 @@
+package com.eureka.servereureka;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ServerEurekaApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
